+++ tests/genes-ts/snapshot/resource-imports/intended/haxe/Rest.ts
@@ -1,0 +1,3 @@
+import {Register} from "../genes/Register.js"
+
+export type NativeRest<T> = T[]